+++ core-ng/src/test/java/core/framework/module/ServerSentEventConfigTest.java
@@ -50,10 +50,4 @@ class ServerSentEventConfigTest {
         assertThat(context).isNotNull();
         assertThat(this.config.context.apiController.beanClasses).contains(TestEvent.class);
     }
-
-    @Test
-    void validate() {
-        config.validate();
-        assertThat(config.context.httpServer.handlerContext.rateControl.hasGroup(ServerSentEventConfig.SSE_CONNECT_GROUP)).isTrue();
-    }
 }